# openapi.json
{"openapi":"3.1.0","info":{"title":"Modular Personal Investment Tracker API","description":"An API to track investments, transactions, and portfolio performance, refactored with SQLModel.","version":"2.0.0"},"paths":{"/investments/":{"get":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Get Investments","description":"List all active investments.","operationId":"get_investments_investments__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InvestmentRead"},"type":"array","title":"Response Get Investments Investments  Get"}}}}}},"post":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Create Investment","description":"Create a new investment record.","operationId":"create_investment_investments__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestmentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investments/{investment_id}/transactions":{"post":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Create Transaction","description":"Log a new transaction (DEPOSIT or WITHDRAWAL).\n\nIf type is WITHDRAWAL for a unit-based asset, this calculates realized P&L\nusing FIFO.","operationId":"create_transaction_investments__investment_id__transactions_post","parameters":[{"name":"investment_id","in":"path","required":true,"schema":{"type":"integer","title":"Investment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investments/{investment_id}/valuation":{"post":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Create Valuation","description":"Record the latest market price/NAV for an investment.","operationId":"create_valuation_investments__investment_id__valuation_post","parameters":[{"name":"investment_id","in":"path","required":true,"schema":{"type":"integer","title":"Investment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValuationCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValuationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investments/{investment_id}/performance":{"get":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Get Performance","description":"Get a full performance summary for a single investment.","operationId":"get_performance_investments__investment_id__performance_get","parameters":[{"name":"investment_id","in":"path","required":true,"schema":{"type":"integer","title":"Investment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerformanceSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/investments/portfolio_summary/":{"get":{"tags":["Investments","Transactions","Valuations","Performance"],"summary":"Get Full Portfolio Summary","description":"Aggregates the performance metrics across all active investments\nto provide a summary of the entire portfolio.","operationId":"get_full_portfolio_summary_investments_portfolio_summary__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioSummaryRead"}}}}}}},"/":{"get":{"tags":["root"],"summary":"Read Root","description":"A simple root endpoint to confirm the API is running.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AssetCategory":{"type":"string","enum":["PF","PPF","FD","MF","STOCK","DERIVATIVES"],"title":"AssetCategory"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InvestmentCreate":{"properties":{"name":{"type":"string","title":"Name"},"symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Symbol"},"asset_category":{"$ref":"#/components/schemas/AssetCategory"},"industry":{"type":"string","title":"Industry"},"risk_category":{"$ref":"#/components/schemas/RiskCategory"},"sip_frequency":{"$ref":"#/components/schemas/SIPFrequency"},"status":{"$ref":"#/components/schemas/Status","default":"ACTIVE"}},"type":"object","required":["name","asset_category","industry","risk_category","sip_frequency"],"title":"InvestmentCreate"},"InvestmentRead":{"properties":{"name":{"type":"string","title":"Name"},"symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Symbol"},"asset_category":{"$ref":"#/components/schemas/AssetCategory"},"industry":{"type":"string","title":"Industry"},"risk_category":{"$ref":"#/components/schemas/RiskCategory"},"sip_frequency":{"$ref":"#/components/schemas/SIPFrequency"},"status":{"$ref":"#/components/schemas/Status","default":"ACTIVE"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","asset_category","industry","risk_category","sip_frequency","id"],"title":"InvestmentRead"},"PerformanceSummary":{"properties":{"investment_name":{"type":"string","title":"Investment Name"},"investment_id":{"type":"integer","title":"Investment Id"},"total_invested_capital":{"type":"number","title":"Total Invested Capital"},"total_withdrawn_capital":{"type":"number","title":"Total Withdrawn Capital"},"total_realized_pnl":{"type":"number","title":"Total Realized Pnl"},"total_units_held":{"type":"number","title":"Total Units Held"},"latest_unit_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latest Unit Price"},"current_market_value":{"type":"number","title":"Current Market Value"},"unrealized_pnl":{"type":"number","title":"Unrealized Pnl"}},"type":"object","required":["investment_name","investment_id","total_invested_capital","total_withdrawn_capital","total_realized_pnl","total_units_held","latest_unit_price","current_market_value","unrealized_pnl"],"title":"PerformanceSummary","description":"Detailed summary for a single investment."},"PortfolioSummaryRead":{"properties":{"investment_count":{"type":"integer","title":"Investment Count"},"total_invested_capital":{"type":"number","title":"Total Invested Capital"},"total_withdrawn_capital":{"type":"number","title":"Total Withdrawn Capital"},"total_realized_pnl":{"type":"number","title":"Total Realized Pnl"},"total_market_value":{"type":"number","title":"Total Market Value"},"total_unrealized_pnl":{"type":"number","title":"Total Unrealized Pnl"},"net_profit_loss":{"type":"number","title":"Net Profit Loss"},"investments":{"items":{"$ref":"#/components/schemas/PerformanceSummary"},"type":"array","title":"Investments"}},"type":"object","required":["investment_count","total_invested_capital","total_withdrawn_capital","total_realized_pnl","total_market_value","total_unrealized_pnl","net_profit_loss","investments"],"title":"PortfolioSummaryRead","description":"Schema for the new endpoint: summary of the entire portfolio."},"RiskCategory":{"type":"string","enum":["LOW","MEDIUM","HIGH","VERY_HIGH"],"title":"RiskCategory"},"SIPFrequency":{"type":"string","enum":["NONE","WEEKLY","MONTHLY","YEARLY"],"title":"SIPFrequency"},"Status":{"type":"string","enum":["ACTIVE","CLOSED"],"title":"Status"},"TransactionCreate":{"properties":{"type":{"$ref":"#/components/schemas/TransactionType"},"date":{"type":"string","format":"date-time","title":"Date"},"cash_amount":{"type":"number","title":"Cash Amount"},"units":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Units"},"price_per_unit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Per Unit"}},"type":"object","required":["type","date","cash_amount"],"title":"TransactionCreate"},"TransactionRead":{"properties":{"type":{"$ref":"#/components/schemas/TransactionType"},"date":{"type":"string","format":"date-time","title":"Date"},"cash_amount":{"type":"number","title":"Cash Amount"},"units":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Units"},"price_per_unit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Per Unit"},"id":{"type":"integer","title":"Id"},"investment_id":{"type":"integer","title":"Investment Id"},"realized_pnl":{"type":"number","title":"Realized Pnl"},"units_remaining":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Units Remaining"}},"type":"object","required":["type","date","cash_amount","id","investment_id","realized_pnl","units_remaining"],"title":"TransactionRead"},"TransactionType":{"type":"string","enum":["DEPOSIT","WITHDRAWAL"],"title":"TransactionType"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValuationCreate":{"properties":{"valuation_date":{"type":"string","format":"date","title":"Valuation Date"},"current_price":{"type":"number","title":"Current Price"}},"type":"object","required":["valuation_date","current_price"],"title":"ValuationCreate"},"ValuationRead":{"properties":{"valuation_date":{"type":"string","format":"date","title":"Valuation Date"},"current_price":{"type":"number","title":"Current Price"},"id":{"type":"integer","title":"Id"},"investment_id":{"type":"integer","title":"Investment Id"}},"type":"object","required":["valuation_date","current_price","id","investment_id"],"title":"ValuationRead"}}}}